+++ tests/opentrons/containers/test_grid.py
@@ -2,7 +2,7 @@ import unittest
 
 from opentrons import containers
 from opentrons import instruments
-from opentrons.robot import Robot
+from opentrons import Robot
 
 
 class GridTestCase(unittest.TestCase):
